+++ src/views/Settings/sections/index.ts
@@ -1,6 +1,5 @@
 import About from "./About";
-import Other from "./Other";
 import Support from "./Support";
 import Themes from "./Themes";
 
-export { About, Other, Support, Themes };
+export { About, Support, Themes };